backend/requirements.txt
PySide6>=6.7
onnxruntime-gpu>=1.18
dxcam>=0.3
numpy>=1.26
opencv-python-headless>=4.9
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.0
websockets>=12.0
mss>=9.0
nvidia-cuda-runtime-cu12
nvidia-cublas-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
